% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esc}
\alias{esc}
\title{Log2 transformed LFQ intensity measurements of the embryonic stem cell (ESC)
proteomes post epiblast-like cell induction}
\format{
Matrix of log2 intensities and sample inforamation in a data frame.
\describe{
\item{esc}{Log2-transformed LFQ intensities. }
\item{smpinfo}{Sample information. }
}
}
\usage{
data(c("esc", "smpinfo"), package = "brainClass")
}
\description{
Processed log2 transformed LFQ intensities of the ESC data set with sample
information.
}
\examples{
data(c("esc", "smpinfo"), package = "brainClass")
esc[1:5, 1:5]
smpinfo
}
\references{
Yang, P., Humphrey, S. J., Cinghu, S., Pathania, R., Oldfield, A. J., Kumar,
D., ... & Jothi, R. (2019). Multi-omic profiling reveals dynamics of the
phased progression of pluripotency. \emph{Cell systems}, 8(5), 427-445.
}
\keyword{datasets}
