% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDA.R
\name{MDA}
\alias{MDA}
\title{Median disagreement (MDA) plot}
\usage{
MDA(raw, normalised, outline = FALSE, ylim = NULL, guides = NULL, title = NULL)
}
\arguments{
\item{raw}{Raw data before normalisation.}

\item{normalised}{Normalised data.}

\item{outline}{Logical. Whether to print the outliers.}

\item{ylim}{The y-axis limits of the output plot.}

\item{guides}{Numeric. Horizontal dashed lines for reference.}

\item{title}{Character. Title of the plot.}
}
\value{
The MDA plot.
}
\description{
For each measurement, calculate the shift in its median abundance level after
normalisation. The \code{MDA} plot is useful when choosing the tuning
parameters.
}
