% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled.sd.R
\name{pooled.sd}
\alias{pooled.sd}
\title{Calculate pooled standard deviations for each protein with missing data}
\usage{
pooled.sd(Y, M)
}
\arguments{
\item{Y}{The m-by-n data matrix with samples on the rows and measurements
(proteins etc. ) on the columns. Missing data (NAs) are allowed.}

\item{M}{The replication structure. See \code{ruv::RUVIII} for details.}
}
\value{
A numeric vector of pooled standard deviation of each measurement.
}
\description{
The \code{pooled.sd} function calculates the pooled standard deviation in
each protein within technical replicates while accommodating possible missing
data. When there is not enough replicates to calculate pooled standard
deviations (number of non-missing samples is less than or equal to number of
groups for example), a simple sample standard deviation is used.
This step is needed when \code{standardise} is enabled in \code{ruvms}.
}
