% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Proteomics data as raw LFQ intensities on the Sydney Heart Bank (SHB) cohort.}
\format{
Matrix of raw LFQ intensities with sample information as a data
frame.
\describe{
\item{raw}{LFQ intensities. }
\item{smpinfo}{Sample information. }
}
}
\usage{
data("shb", package = "ruvms")
}
\description{
LFQ intensities of the SHB proteomics data set with sample information.
}
\examples{
data("shb", package = "ruvms")
raw[1:5, 1:5]
head(smpinfo)
}
\references{
Li, M., Parker, B. L., Pearson, E., Hunter, B., Cao, J., Koay, Y. C., ... &
O’Sullivan, J. F. (2020). Core functional nodes and sex-specific pathways in
human ischaemic and dilated cardiomyopathy. \emph{Nature communications},
11(1), 1-12.
}
\keyword{datasets}
