% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLA.R
\name{RLA}
\alias{RLA}
\title{Relative log abundance (RLA) plot}
\usage{
RLA(
  data,
  smpID = NULL,
  repCol = NULL,
  repLabel = NULL,
  smpName = NULL,
  ylim = NULL,
  guides = NULL,
  title = NULL
)
}
\arguments{
\item{data}{The m-by-n data matrix with samples on the rows and measurements (
proteins etc.) on the columns. Missing data (NAs) are allowed.}

\item{smpID}{Sample IDs of the data. Default is \code{NULL}. This argument is
useful when one wishes to display the samples in an order different from the
sample orders of input data.}

\item{repCol}{The replication color codes. Useful when one wishes to colour
the RLA boxes to indicate replications. Default is \code{NULL}.}

\item{repLabel}{Color legend names.}

\item{smpName}{Sample names. Sample names to display on the axis.
Default is \code{NULL}.}

\item{ylim}{The y-axis labels. A numeric vector of length 2.}

\item{guides}{Numeric. Horizontal dashed lines for reference.}

\item{title}{Character. Title of the output plot.}
}
\value{
The RLA plot.
}
\description{
The \code{RLA} function is a wrapper function to help easily visualise
normalisation results on a data set. It is equivalent to the reltaive log
expression (RLE) plot under the gene expression contexts.
}
\references{
Gandolfo, L. C., & Speed, T. P. (2018). RLE plots: Visualizing unwanted
variation in high dimensional data. \emph{PloS one}, 13(2), e0191629.
}
